Finished implementing test suite for gencore

+++ genstone/gencore/genprocess.c
@@ -13,6 +13,7 @@ GEN_PRAGMA(GEN_PRAGMA_DIAGNOSTIC_REGION_IGNORE("-Weverything"))
 #include <sys/fcntl.h>
 #include <sys/wait.h>
 #include <signal.h>
+#include <errno.h>
 #endif
 GEN_PRAGMA(GEN_PRAGMA_DIAGNOSTIC_REGION_END)
 
@@ -44,6 +45,8 @@ static void gen_process_internal_create_with_redirect_cleanup_arguments_copy(cha
     }
 }
 
+// TODO: Take environment lengths and argument lengths
+// TODO: Take const-qualified environment
 gen_error_t* gen_process_create_with_redirect(const char* const restrict executable_path, const size_t executable_path_length, const char* const* const restrict arguments, const size_t arguments_length, char** const restrict environment, const size_t environment_length, gen_filesystem_handle_t* const restrict filesystem_handle, gen_process_t* const restrict out_process) {
 	GEN_TOOLING_AUTO gen_error_t* error = gen_tooling_push(GEN_FUNCTION_NAME, (void*) gen_filesystem_path_canonicalize, GEN_FILE_NAME);
 	if(error) return error;
@@ -132,16 +135,40 @@ gen_error_t* gen_process_kill(const gen_process_t* const restrict process) {
 
 #if GEN_PLATFORM == GEN_LINUX || GEN_PLATFORM == GEN_OSX
     int result = kill(*process, SIGKILL);
+    if(result == -1 && errno == ESRCH) return NULL;
     if(result == -1) return gen_error_attach_backtrace_formatted(gen_error_type_from_errno(), GEN_LINE_NUMBER, "Could not kill process %uz: %t", *process, gen_error_description_from_errno());
 #endif
 
     return NULL;
 }
 
+gen_error_t* gen_process_check(const gen_process_t* const restrict process, bool* const restrict out_alive) {
+	GEN_TOOLING_AUTO gen_error_t* error = gen_tooling_push(GEN_FUNCTION_NAME, (void*) gen_process_check, GEN_FILE_NAME);
+	if(error) return error;
+
+	if(!process) return gen_error_attach_backtrace(GEN_ERROR_INVALID_PARAMETER, GEN_LINE_NUMBER, "`process` was `NULL`");
+	if(!out_alive) return gen_error_attach_backtrace(GEN_ERROR_INVALID_PARAMETER, GEN_LINE_NUMBER, "`out_alive` was `NULL`");
+
+#if GEN_PLATFORM == GEN_LINUX || GEN_PLATFORM == GEN_OSX
+    int result = kill(*process, 0);
+    if(result == -1 && errno == ESRCH) {
+        *out_alive = false;
+        return NULL;
+    }
+    else if(result == -1) return gen_error_attach_backtrace_formatted(gen_error_type_from_errno(), GEN_LINE_NUMBER, "Could not check process %uz: %t", *process, gen_error_description_from_errno());
+
+    *out_alive = true;
+#endif
+
+    return NULL;
+}
+
+// TODO: Add "Force Unix" mode everywhere - not just genfilesystem
 #if GEN_PLATFORM == GEN_LINUX || GEN_PLATFORM == GEN_OSX
 extern char** environ;
 #endif
 
+// TODO: Make this copy-out - output length, then lengths, then the environment 
 gen_error_t* gen_process_get_environment(char*** const restrict out_environment, size_t* const restrict out_length) {
 	GEN_TOOLING_AUTO gen_error_t* error = gen_tooling_push(GEN_FUNCTION_NAME, (void*) gen_process_kill, GEN_FILE_NAME);
 	if(error) return error;
